server/mqtt/requirements.txt
filelock==3.16.1
fsspec==2024.10.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.24.0
paho-mqtt==1.6.1
pillow==11.0.0
sympy==1.13.3
torch
torchvision
typing_extensions==4.12.2
confluent-kafka==2.6.0